+++ noxfile.py
@@ -22,8 +22,8 @@ except ImportError:
     raise SystemExit(dedent(message)) from None
 
 
-package = "pydatautils"
-python_versions = ["3.10", "3.9", "3.8", "3.7"]
+package = "pydataassist"
+python_versions = ["3.10", "3.9"]
 nox.needs_version = ">= 2021.6.6"
 nox.options.sessions = (
     "pre-commit",
