+++ src/pydataassist/__main__.py
@@ -5,8 +5,8 @@ import click
 @click.command()
 @click.version_option()
 def main() -> None:
-    """PyDataUtils."""
+    """PyDataAssist."""
 
 
 if __name__ == "__main__":
-    main(prog_name="pydatautils")  # pragma: no cover
+    main(prog_name="pydataassist")  # pragma: no cover
